refactor(engine): split the signoff collector out of the package init
The package __init__ carried the full 1305-line collector. Move the
public data model into models.py, artifact discovery into discovery.py,
analysis refresh and issue validation into analysis.py, and keep only
orchestration, materialization, and archiving in collector.py; the
package __init__ is now just the public re-exports.
